[refactor] Refactor file architecture - Add path 'social' - Seperate token package
