+++ src/pkg/server/auth/auth.go
@@ -19,10 +19,10 @@ package auth
 import (
 	"github.com/110billion/usermanagerservice/src/internal/apiserver"
 	"github.com/110billion/usermanagerservice/src/internal/wrapper"
-	"github.com/110billion/usermanagerservice/src/pkg/server/auth/facebook"
-	"github.com/110billion/usermanagerservice/src/pkg/server/auth/google"
 	"github.com/110billion/usermanagerservice/src/pkg/server/auth/login"
 	"github.com/110billion/usermanagerservice/src/pkg/server/auth/signup"
+	"github.com/110billion/usermanagerservice/src/pkg/server/auth/social/facebook"
+	"github.com/110billion/usermanagerservice/src/pkg/server/auth/social/google"
 	"github.com/go-logr/logr"
 )
