+++ src/pkg/server/auth/signup/signup.go
@@ -1,3 +1,19 @@
+/*
+ Copyright 2021 The 110 billion Authors
+
+ Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+
+     http://www.apache.org/licenses/LICENSE-2.0
+
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+*/
+
 package signup
 
 import (
@@ -6,7 +22,7 @@ import (
 	"github.com/110billion/usermanagerservice/src/internal/apiserver"
 	"github.com/110billion/usermanagerservice/src/internal/utils"
 	"github.com/110billion/usermanagerservice/src/internal/wrapper"
-	"github.com/110billion/usermanagerservice/src/pkg/server/database"
+	"github.com/110billion/usermanagerservice/src/pkg/database"
 	"github.com/go-logr/logr"
 	_ "github.com/lib/pq"
 	"golang.org/x/crypto/bcrypt"
