+++ src/pkg/server/auth/token/token.go
@@ -1,0 +1,46 @@
+/*
+ Copyright 2021 The 110 billion Authors
+
+ Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+
+     http://www.apache.org/licenses/LICENSE-2.0
+
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+*/
+
+package token
+
+import (
+	"github.com/dgrijalva/jwt-go"
+	"os"
+	"time"
+)
+
+type claims struct {
+	userEmail string
+	jwt.StandardClaims
+}
+
+func GetJwtToken(email string) (string, error) {
+	expirationTime := time.Now().Add(5 * time.Hour)
+	jwtKey := os.Getenv("JWT_SECRET_KEY")
+	claims := &claims{
+		userEmail: email,
+		StandardClaims: jwt.StandardClaims{
+			ExpiresAt: expirationTime.Unix(),
+		},
+	}
+
+	token := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
+	tokenString, err := token.SignedString([]byte(jwtKey))
+	if err != nil {
+		return "", err
+	}
+	return tokenString, nil
+}
