+++ src/pkg/server/server.go
@@ -22,8 +22,8 @@ import (
 	"github.com/110billion/usermanagerservice/src/internal/utils"
 	"github.com/110billion/usermanagerservice/src/internal/wrapper"
 	"github.com/110billion/usermanagerservice/src/pkg/server/auth"
-	"github.com/110billion/usermanagerservice/src/pkg/server/auth/facebook"
-	"github.com/110billion/usermanagerservice/src/pkg/server/auth/google"
+	"github.com/110billion/usermanagerservice/src/pkg/server/auth/social/facebook"
+	"github.com/110billion/usermanagerservice/src/pkg/server/auth/social/google"
 	"github.com/gorilla/mux"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"net/http"
